refactor: entity 위치 변경
- 각 도메인 -> common/entity
