+++ src/main/java/com/example/budongbudong/common/entity/PropertyImage.java
@@ -1,7 +1,5 @@
-package com.example.budongbudong.domain.propertyimage.entity;
+package com.example.budongbudong.common.entity;
 
-import com.example.budongbudong.common.entity.BaseEntity;
-import com.example.budongbudong.domain.property.entity.Property;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
@@ -9,7 +7,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
-@Table(name="property_images")
+@Table(name = "property_images")
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class PropertyImage extends BaseEntity {
@@ -19,10 +17,10 @@ public class PropertyImage extends BaseEntity {
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name="property_id",nullable = false)
+    @JoinColumn(name = "property_id", nullable = false)
     private Property property;
 
-    @Column(name = "image_url",nullable = false,length = 255)
+    @Column(name = "image_url", nullable = false, length = 255)
     private String imageUrl;
 
     @Builder
