+++ src/main/java/com/example/budongbudong/domain/auction/repository/AuctionRepository.java
@@ -1,8 +1,8 @@
 package com.example.budongbudong.domain.auction.repository;
 
+import com.example.budongbudong.common.entity.Auction;
 import com.example.budongbudong.common.exception.CustomException;
 import com.example.budongbudong.common.exception.ErrorCode;
-import com.example.budongbudong.domain.auction.entity.Auction;
 import com.example.budongbudong.domain.auction.enums.AuctionStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
