+++ src/main/java/com/example/budongbudong/domain/auctionwinner/repository/AuctionWinnerRepository.java
@@ -1,6 +1,6 @@
 package com.example.budongbudong.domain.auctionwinner.repository;
 
-import com.example.budongbudong.domain.auctionwinner.entity.AuctionWinner;
+import com.example.budongbudong.common.entity.AuctionWinner;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AuctionWinnerRepository extends JpaRepository<AuctionWinner, Long> {
