+++ src/main/java/com/example/budongbudong/domain/bid/dto/response/CreateBidResponse.java
@@ -1,6 +1,6 @@
 package com.example.budongbudong.domain.bid.dto.response;
 
-import com.example.budongbudong.domain.bid.entity.Bid;
+import com.example.budongbudong.common.entity.Bid;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
