+++ src/main/java/com/example/budongbudong/domain/bid/repository/BidRepository.java
@@ -1,9 +1,9 @@
 package com.example.budongbudong.domain.bid.repository;
 
+import com.example.budongbudong.common.entity.Auction;
+import com.example.budongbudong.common.entity.Bid;
 import com.example.budongbudong.common.exception.CustomException;
 import com.example.budongbudong.common.exception.ErrorCode;
-import com.example.budongbudong.domain.auction.entity.Auction;
-import com.example.budongbudong.domain.bid.entity.Bid;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
