+++ src/main/java/com/example/budongbudong/domain/property/service/PropertyService.java
@@ -1,25 +1,25 @@
 package com.example.budongbudong.domain.property.service;
 
-import com.example.budongbudong.domain.property.client.AptItem;
-import com.example.budongbudong.domain.property.client.AptMapper;
-import com.example.budongbudong.domain.property.client.AptResponse;
+import com.example.budongbudong.common.entity.Auction;
+import com.example.budongbudong.common.entity.Property;
+import com.example.budongbudong.common.entity.User;
 import com.example.budongbudong.common.exception.CustomException;
 import com.example.budongbudong.common.exception.ErrorCode;
 import com.example.budongbudong.common.response.CustomPageResponse;
-import com.example.budongbudong.domain.property.client.AptClient;
-import com.example.budongbudong.domain.property.dto.request.CreatePropertyRequest;
-import com.example.budongbudong.domain.property.dto.request.UpdatePropertyRequest;
 import com.example.budongbudong.domain.auction.dto.response.AuctionResponse;
-import com.example.budongbudong.domain.auction.entity.Auction;
 import com.example.budongbudong.domain.auction.enums.AuctionStatus;
 import com.example.budongbudong.domain.auction.repository.AuctionRepository;
+import com.example.budongbudong.domain.property.client.AptClient;
+import com.example.budongbudong.domain.property.client.AptItem;
+import com.example.budongbudong.domain.property.client.AptMapper;
+import com.example.budongbudong.domain.property.client.AptResponse;
+import com.example.budongbudong.domain.property.dto.request.CreatePropertyRequest;
+import com.example.budongbudong.domain.property.dto.request.UpdatePropertyRequest;
 import com.example.budongbudong.domain.property.dto.response.CreateApiResponse;
 import com.example.budongbudong.domain.property.dto.response.ReadAllPropertyResponse;
 import com.example.budongbudong.domain.property.dto.response.ReadPropertyResponse;
-import com.example.budongbudong.domain.property.entity.Property;
 import com.example.budongbudong.domain.property.repository.PropertyRepository;
 import com.example.budongbudong.domain.propertyimage.service.PropertyImageService;
-import com.example.budongbudong.domain.user.entity.User;
 import com.example.budongbudong.domain.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
