+++ src/main/java/com/example/budongbudong/domain/propertyimage/repository/PropertyImageRepository.java
@@ -1,6 +1,6 @@
 package com.example.budongbudong.domain.propertyimage.repository;
 
-import com.example.budongbudong.domain.propertyimage.entity.PropertyImage;
+import com.example.budongbudong.common.entity.PropertyImage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PropertyImageRepository extends JpaRepository<PropertyImage, Long> {
