+++ src/main/java/com/example/budongbudong/domain/user/repository/UserRepository.java
@@ -1,13 +1,13 @@
 package com.example.budongbudong.domain.user.repository;
 
+import com.example.budongbudong.common.entity.User;
 import com.example.budongbudong.common.exception.CustomException;
 import com.example.budongbudong.common.exception.ErrorCode;
-import com.example.budongbudong.domain.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User,Long> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
     boolean existsByEmail(String userEmail);
